requirements.txt
psycopg2-binary==2.8.5
Scrapy==2.3.0
scrapy-rotating-proxies==0.6.2
slackclient==2.8.0
SQLAlchemy==1.3.18
python-dotenv
scrapyscript
crochet